requirements.txt
alembic==0.9.9
aniso8601==3.0.0
click==6.7
Flask-CORS==3.0.3
Flask-HTTPAuth==3.2.3
Flask-JWT-Extended==3.9.1
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
Flask==1.0.1
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
passlib==1.7.1
psycopg2==2.7.4
PyJWT==1.6.3
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.4
six==1.11.0
SQLAlchemy==1.2.6
Werkzeug==0.14.1
